Moved RestApiClientBase to stratis.bitcoin to allow client implementations in features
