+++ src/Stratis.Features.FederatedPeg/Stratis.Features.FederatedPeg.csproj
@@ -15,8 +15,6 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.1.3" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="2.1.1" />
-    <PackageReference Include="Polly" Version="6.1.2" />
     <PackageReference Include="Tracer.Fody" Version="2.2.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
